chore(acp176): change MinTargetPerSecond to be a variable - Needed for load testing on the avalanchego side - Update associated constants to be variables

+++ core/state_processor_test.go
@@ -156,7 +156,7 @@ func TestStateProcessorErrors(t *testing.T) {
 			{ // ErrGasLimitReached
 				txs: []*types.Transaction{
 					// This test was modified to account for ACP-176 gas limits
-					makeTx(key1, 0, common.Address{}, big.NewInt(0), acp176.MinMaxCapacity+1, big.NewInt(acp176.MinGasPrice), nil),
+					makeTx(key1, 0, common.Address{}, big.NewInt(0), uint64(acp176.MinMaxCapacity+1), big.NewInt(acp176.MinGasPrice), nil),
 				},
 				want: "could not apply tx 0 [0xb709565c056a68a4b4dc7714ae901a0f03663bb98f9fa58a10b88ec614362caf]: gas limit reached",
 			},
